% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{addition}
\alias{addition}
\title{outputs the sum of two arguments}
\usage{
addition(arg1, arg2)
}
\arguments{
\item{arg1}{numeric first argument}

\item{arg2}{numeric second argument}
}
\value{
numeric Sum
}
\description{
outputs the sum of two arguments
the inputs should be numeric
}
